@@ -193,21 +193,19 @@ export const getDeploymentsByFilter: ResolverFn = async (
   }
 
   let queryBuilder = knex.select("deployment.*").from('deployment').
-      join('environment', 'deployment.environment', '=', 'environment.id').
-      join('project', 'environment.project', '=', 'project.id').
-      where('environment.deleted', '=', '0000-00-00 00:00:00');
+      join('environment', 'deployment.environment', '=', 'environment.id');
+
+  if (userProjectIds) {
+      queryBuilder = queryBuilder.whereIn('environment.project', userProjectIds);
+  }
 
   if(openshifts) {
     queryBuilder = queryBuilder.whereIn('environment.openshift', openshifts);
   }
 
-
   queryBuilder = queryBuilder.whereIn('deployment.status', deploymentStatus);
 
-
-  if (userProjectIds) {
-    queryBuilder = queryBuilder.whereIn('project.id', userProjectIds);
-  }
+  queryBuilder = queryBuilder.where('environment.deleted', '=', '0000-00-00 00:00:00');
 
   const queryBuilderString = queryBuilder.toString();
 
@@ -1249,7 +1247,7 @@ export const switchActiveStandby: ResolverFn = async (
     data.task.id = sourceTaskData.addTask.id.toString();
 
     // queue the task to trigger the migration
-    await createMiscTask({ key: 'route:migrate', data });
+    await createMiscTask({ key: 'task:activestandby', data });
 
     // return the task id and remote id
     var retData = {
