@@ -128,14 +128,17 @@ export const Sql = {
         created
       })
       .toString(),
-  deleteProblem: (environment, identifier) =>
-    knex('environment_problem')
+  deleteProblem: (environment, identifier, service) => {
+    let q = knex('environment_problem')
       .where({
         environment: environment,
         identifier: identifier
-      })
-      .del()
-      .toString(),
+      });
+    if (service != undefined) {
+      q.where('lagoon_service', service);
+    }
+    return q.del().toString();
+  },
   deleteProblemsFromSource: (environment, source, service) =>
     knex('environment_problem')
       .where({
