@@ -307,6 +307,7 @@ const typeDefs = gql`
   input DeleteProblemInput {
     environment: Int!
     identifier: String!
+    service: String
   }
 
   input DeleteProblemsFromSourceInput {
@@ -473,6 +474,8 @@ const typeDefs = gql`
     id: String
     name: String
     role: GroupRole
+    groupType: String
+    organization: Int
   }
 
   interface GroupInterface {
@@ -481,7 +484,9 @@ const typeDefs = gql`
     type: String
     groups: [GroupInterface]
     members: [GroupMembership]
+    memberCount: Int
     projects: [Project]
+    organization: Int
   }
 
   type Group implements GroupInterface {
@@ -490,7 +495,31 @@ const typeDefs = gql`
     type: String
     groups: [GroupInterface]
     members: [GroupMembership]
+    memberCount: Int
     projects: [Project]
+    organization: Int
+  }
+
+  interface OrgGroupInterface {
+    id: String
+    name: String
+    type: String
+    groups: [OrgGroupInterface]
+    members: [GroupMembership]
+    memberCount: Int
+    projects: [OrgProject]
+    organization: Int
+  }
+
+  type OrgGroup implements OrgGroupInterface {
+    id: String
+    name: String
+    type: String
+    groups: [OrgGroupInterface]
+    members: [GroupMembership]
+    memberCount: Int
+    projects: [OrgProject]
+    organization: Int
   }
 
   type Openshift {
@@ -595,6 +624,10 @@ const typeDefs = gql`
     Name of project
     """
     name: String
+    """
+    ID of organization
+    """
+    organization: Int
     """
     Git URL, needs to be SSH Git URL in one of these two formats
     - git@172.17.0.1/project1.git
@@ -934,6 +967,10 @@ const typeDefs = gql`
     backupId: String
     status: String
     restoreLocation: String
+    """
+    The size of the restored file in bytes
+    """
+    restoreSize: Int
     created: String
   }
 
@@ -1022,6 +1059,107 @@ const typeDefs = gql`
     environments: [Environment]
   }
 
+  type OrgUser {
+    id: String
+    email: String
+    firstName: String
+    lastName: String
+    owner: Boolean
+    comment: String
+    groupRoles: [GroupRoleInterface]
+  }
+
+  type Organization {
+    id: Int
+    name: String
+    friendlyName: String
+    description: String
+    quotaProject: Int
+    quotaGroup: Int
+    quotaNotification: Int
+    quotaEnvironment: Int
+    quotaRoute: Int
+    deployTargets: [Openshift]
+    projects: [OrgProject]
+    environments: [OrgEnvironment]
+    groups: [OrgGroupInterface]
+    owners: [OrgUser]
+    notifications(type: NotificationType): [Notification]
+  }
+
+  input AddOrganizationInput {
+    id: Int
+    name: String!
+    friendlyName: String
+    description: String
+    quotaProject: Int
+    quotaGroup: Int
+    quotaNotification: Int
+    quotaEnvironment: Int
+    quotaRoute: Int
+  }
+
+  input DeleteOrganizationInput {
+    id: Int!
+  }
+
+  input UpdateOrganizationPatchInput {
+    name: String
+    friendlyName: String
+    description: String
+    quotaProject: Int
+    quotaGroup: Int
+    quotaNotification: Int
+    quotaEnvironment: Int
+    quotaRoute: Int
+  }
+
+  input UpdateOrganizationInput {
+    id: Int!
+    patch: UpdateOrganizationPatchInput!
+  }
+
+  """
+  OrgProject is a small selection of fields for organization owners to view
+  """
+  type OrgProject {
+    id: Int
+    name: String
+    organization: Int
+    groups: [OrgGroupInterface]
+    groupCount: Int
+    notifications: [OrganizationNotification]
+  }
+
+  """
+  OrgEnvironment is a small selection of fields for organization owners to view
+  """
+  type OrgEnvironment {
+    id: Int
+    name: String
+    project: OrgProject
+    deployType: String
+    deployHeadRef: String
+    deployTitle: String
+    autoIdle: Int
+    environmentType: String
+    openshiftProjectName: String
+    kubernetesNamespaceName: String
+    updated: String
+    created: String
+    deleted: String
+    route: String
+    routes: String
+    services: [EnvironmentService]
+    openshift: Openshift
+    kubernetes: Kubernetes
+  }
+
+  type OrganizationNotification {
+    name: String
+    type: NotificationType
+  }
+
   type DeployTargetConfig {
     id: Int
     project: Project
@@ -1082,6 +1220,11 @@ const typeDefs = gql`
     key: String!
   }
 
+  input ProjectOrgGroupsInput {
+    project: Int!
+    organization: Int!
+  }
+
   input EnvVariableByProjectEnvironmentNameInput {
     environment: String
     project: String!
@@ -1129,6 +1272,7 @@ const typeDefs = gql`
     Returns Project Object by a given name
     """
     projectByName(name: String!): Project
+    orgProjectByName(name: String!): OrgProject
     """
     Returns all Environment Objects for a specified Kubernetes matching given filter (all if no filter defined)
     """
@@ -1137,6 +1281,16 @@ const typeDefs = gql`
     Returns Group Object by a given name
     """
     groupByName(name: String!): GroupInterface
+    groupByNameAndOrganization(name: String!, organization: Int!): OrgGroupInterface
+    """
+    Retrieves all users that have been added to groups within an organization.
+    """
+    usersByOrganization(organization: Int!): [OrgUser]
+    """
+    Retrieve information about a specific user within groups within an organization.
+    This will only return group information if this user is in any groups within this organization
+    """
+    userByEmailAndOrganization(email: String!, organization: Int!): OrgUser
     """
     Returns Project Object by a given gitUrl (only the first one if there are multiple)
     """
@@ -1182,7 +1336,7 @@ const typeDefs = gql`
     """
     Returns all Project Objects matching given filters (all if no filter defined)
     """
-    allProjects(createdAfter: String, gitUrl: String, order: ProjectOrderType): [Project]
+    allProjects(createdAfter: String, gitUrl: String, order: ProjectOrderType, buildImage: Boolean): [Project]
     """
     Returns all Project Objects matching metadata filters
     """
@@ -1190,11 +1344,11 @@ const typeDefs = gql`
     """
     Returns all OpenShift Objects
     """
-    allOpenshifts(disabled: Boolean): [Openshift]
+    allOpenshifts(disabled: Boolean, buildImage: Boolean): [Openshift]
     """
     Returns all Kubernetes Objects
     """
-    allKubernetes(disabled: Boolean): [Kubernetes]
+    allKubernetes(disabled: Boolean, buildImage: Boolean): [Kubernetes]
     """
     Returns all Environments matching given filter (all if no filter defined)
     """
@@ -1259,7 +1413,26 @@ const typeDefs = gql`
     """
     deployTargetConfigsByDeployTarget(deployTarget: Int!) : [DeployTargetConfig]  @deprecated(reason: "Unstable API, subject to breaking changes in any release. Use at your own risk")
     allDeployTargetConfigs: [DeployTargetConfig]  @deprecated(reason: "Unstable API, subject to breaking changes in any release. Use at your own risk")
+    """
+    List all organizations
+    """
+    allOrganizations: [Organization] @deprecated(reason: "Unstable API, subject to breaking changes in any release. Use at your own risk")
+    """
+    Get an organization by its ID
+    """
+    organizationById(id: Int!): Organization
+    organizationByName(name: String!): Organization
+    getGroupProjectOrganizationAssociation(input: AddGroupToOrganizationInput!): String  @deprecated(reason: "Use checkBulkImportProjectsAndGroupsToOrganization instead")
+    getProjectGroupOrganizationAssociation(input: ProjectOrgGroupsInput!): String  @deprecated(reason: "Use checkBulkImportProjectsAndGroupsToOrganization instead")
     getEnvVariablesByProjectEnvironmentName(input: EnvVariableByProjectEnvironmentNameInput!): [EnvKeyValue]
+    checkBulkImportProjectsAndGroupsToOrganization(input: AddProjectToOrganizationInput!): ProjectGroupsToOrganization
+  }
+
+  type ProjectGroupsToOrganization {
+    projects: [Project]
+    groups: [GroupInterface]
+    otherOrgProjects: [Project]
+    otherOrgGroups: [GroupInterface]
   }
 
   # Must provide id OR name
@@ -1331,6 +1504,8 @@ const typeDefs = gql`
     productionBuildPriority: Int
     developmentBuildPriority: Int
     deploymentsDisabled: Int
+    organization: Int
+    addOrgOwner: Boolean
     buildImage: String
     sharedBaasBucket: Boolean
   }
@@ -1608,27 +1783,32 @@ const typeDefs = gql`
   input AddNotificationMicrosoftTeamsInput {
     name: String!
     webhook: String!
+    organization: Int
   }
   input AddNotificationEmailInput {
     name: String!
     emailAddress: String!
+    organization: Int
   }
 
   input AddNotificationRocketChatInput {
     name: String!
     webhook: String!
     channel: String!
+    organization: Int
   }
 
   input AddNotificationWebhookInput {
     name: String!
     webhook: String!
+    organization: Int
   }
 
   input AddNotificationSlackInput {
     name: String!
     webhook: String!
     channel: String!
+    organization: Int
   }
 
   input DeleteNotificationMicrosoftTeamsInput {
@@ -1670,6 +1850,7 @@ const typeDefs = gql`
     lastName: String
     comment: String
     gitlabId: Int
+    resetPassword: Boolean
   }
 
   input UpdateUserPatchInput {
@@ -1686,6 +1867,16 @@ const typeDefs = gql`
   }
 
   input DeleteUserInput {
+    user: UserInput!
+  }
+
+  input addUserToOrganizationInput {
+    user: UserInput!
+    organization: Int!
+    owner: Boolean
+  }
+
+  input ResetUserPasswordInput {
     user: UserInput!
   }
 
@@ -1732,6 +1923,26 @@ const typeDefs = gql`
   input UpdateProjectInput {
     id: Int!
     patch: UpdateProjectPatchInput!
+  }
+
+  input AddProjectToOrganizationInput {
+    project: Int!
+    organization: Int!
+  }
+
+  input RemoveProjectFromOrganizationInput {
+    project: Int!
+    organization: Int!
+  }
+
+  input AddDeployTargetToOrganizationInput {
+    deployTarget: Int!
+    organization: Int!
+  }
+
+  input RemoveDeployTargetFromOrganizationInput {
+    deployTarget: Int!
+    organization: Int!
   }
 
   input UpdateOpenshiftPatchInput {
@@ -1980,6 +2191,13 @@ const typeDefs = gql`
     parentGroup: GroupInput
   }
 
+  input AddGroupToOrganizationInput {
+    name: String!
+    organization: Int!
+    parentGroup: GroupInput
+    addOrgOwner: Boolean
+  }
+
   input UpdateGroupPatchInput {
     name: String
   }
@@ -2012,6 +2230,11 @@ const typeDefs = gql`
   input ProjectGroupsInput {
     project: ProjectInput!
     groups: [GroupInput!]!
+  }
+
+  input UserOrganizationInput {
+    user: UserInput!
+    organization: Int!
   }
 
   input BulkDeploymentLatestInput {
@@ -2108,6 +2331,17 @@ const typeDefs = gql`
     removeAllSshKeysFromAllUsers: String
     addUser(input: AddUserInput!): User
     updateUser(input: UpdateUserInput!): User
+    """
+    Add a user to an organization as a viewer or owner of the organization.
+    This allows the user to view or manage the organizations groups, projects, and notifications
+    """
+    addUserToOrganization(input: addUserToOrganizationInput!): Organization
+    """
+    Remove a viewer or owner from an organization.
+    This removes the users ability to view or manage the organizations groups, projects, and notifications
+    """
+    removeUserFromOrganization(input: addUserToOrganizationInput!): Organization
+    resetUserPassword(input: ResetUserPasswordInput!): String
     deleteUser(input: DeleteUserInput!): String
     deleteAllUsers: String
     addDeployment(input: AddDeploymentInput!): Deployment
@@ -2177,6 +2411,11 @@ const typeDefs = gql`
     removeUserFromGroup(input: UserGroupInput!): GroupInterface
     addGroupsToProject(input: ProjectGroupsInput): Project
     removeGroupsFromProject(input: ProjectGroupsInput!): Project
+    """
+    This is a way to quickly remove a user from all groups within an organization.
+    Effectively removing all access to any projects that are assigned to those groups.
+    """
+    removeUserFromOrganizationGroups(input: UserOrganizationInput): Organization
     updateProjectMetadata(input: UpdateMetadataInput!): Project
     removeProjectMetadataByKey(input: RemoveMetadataInput!): Project
     addDeployTargetConfig(input: AddDeployTargetConfigInput!): DeployTargetConfig  @deprecated(reason: "Unstable API, subject to breaking changes in any release. Use at your own risk")
@@ -2184,6 +2423,49 @@ const typeDefs = gql`
     deleteDeployTargetConfig(input: DeleteDeployTargetConfigInput!): String  @deprecated(reason: "Unstable API, subject to breaking changes in any release. Use at your own risk")
     deleteAllDeployTargetConfigs: String  @deprecated(reason: "Unstable API, subject to breaking changes in any release. Use at your own risk")
     updateEnvironmentDeployTarget(environment: Int!, deployTarget: Int!): Environment
+    """
+    Add an organization
+    """
+    addOrganization(input: AddOrganizationInput!): Organization
+    """
+    Update an organization
+    """
+    updateOrganization(input: UpdateOrganizationInput!): Organization
+    """
+    Delete an organization
+    """
+    deleteOrganization(input: DeleteOrganizationInput!): String
+    """
+    Add a group to an organization
+    """
+    addGroupToOrganization(input: AddGroupToOrganizationInput!): OrgGroupInterface  @deprecated(reason: "Use bulkImportProjectsAndGroupsToOrganization instead")
+    """
+    Add an existing group to an organization
+    """
+    addExistingGroupToOrganization(input: AddGroupToOrganizationInput!): OrgGroupInterface  @deprecated(reason: "Use bulkImportProjectsAndGroupsToOrganization instead")
+    """
+    Add an existing project to an organization
+    """
+    addExistingProjectToOrganization(input: AddProjectToOrganizationInput): Project  @deprecated(reason: "Use bulkImportProjectsAndGroupsToOrganization instead")
+    """
+    Remove a project from an organization, this will return the project to a state where it has no groups or notifications associated to it
+    """
+    removeProjectFromOrganization(input: RemoveProjectFromOrganizationInput): Project
+    """
+    Add a deploytarget to an organization
+    """
+    addDeployTargetToOrganization(input: AddDeployTargetToOrganizationInput): Organization
+    """
+    Remove a deploytarget from an organization
+    """
+    removeDeployTargetFromOrganization(input: RemoveDeployTargetFromOrganizationInput): Organization
+    """
+    Run the query checkBulkImportProjectsAndGroupsToOrganization first to see the changes that would be made before executing this, as it may contain undesirable changes
+    Add an existing project to an organization, this will include all the groups and all the projects that those groups contain
+    Optionally detach any notifications attached to the projects, they will be need to be recreated within the organization afterwards
+    This mutation performs a lot of actions, on big project and group imports, if it times out, subsequent runs will perform only the changes necessary
+    """
+    bulkImportProjectsAndGroupsToOrganization(input: AddProjectToOrganizationInput, detachNotification: Boolean): ProjectGroupsToOrganization
   }
 
   type Subscription {
